. objbase.cpp, fuseif.h, defines.h: moved shared lock from fuseif.h to defines.h, so that it can be used by other modules. . propids.h: added a new property propStmHandle.
