+++ fuse/fuseif.h
@@ -82,200 +82,7 @@ gint32 AddSvcPoint(
 const char* IfStateToString(
     EnumIfState dwState );
 
-class CSharedLock
-{
-    stdmutex m_oLock;
-    bool m_bWrite = false;
-    gint32 m_iReadCount = 0;
-    sem_t m_semReader;
-    sem_t m_semWriter;
-    // true : reader, false: writer
-    std::deque< bool > m_queue;
-
-    public:
-    CSharedLock()
-    {
-        Sem_Init( &m_semReader, 0, 0 );
-        Sem_Init( &m_semWriter, 0, 0 );
-    }
-
-    ~CSharedLock()
-    {
-        sem_destroy( &m_semReader );
-        sem_destroy( &m_semWriter );
-    }
-
-    inline gint32 LockRead()
-    {
-        CStdMutex oLock( m_oLock );
-        if( m_bWrite )
-        {
-            m_queue.push_back( true );
-            oLock.Unlock();
-            return Sem_Wait_Wakable( &m_semReader );
-        }
-        else
-        {
-            if( m_queue.empty() )
-            {
-                m_iReadCount++;
-                oLock.Unlock();
-                return STATUS_SUCCESS;
-            }
-            m_queue.push_back( true );
-            oLock.Unlock();
-            return Sem_Wait_Wakable( &m_semReader );
-        }
-    }
-
-    inline gint32 LockWrite()
-    {
-        CStdMutex oLock( m_oLock );
-        if( m_bWrite )
-        {
-            m_queue.push_back( false );
-            oLock.Unlock();
-            return Sem_Wait_Wakable( &m_semWriter );
-        }
-        else if( m_iReadCount > 0 )
-        {
-            m_queue.push_back( false );
-            oLock.Unlock();
-            return Sem_Wait_Wakable( &m_semWriter );
-        }
-        else if( m_iReadCount == 0 )
-        {
-            m_bWrite = true;
-            return 0;
-        }
-        return -EFAULT;
-    }
-
-    inline gint32 ReleaseRead()
-    {
-        CStdMutex oLock( m_oLock );
-        --m_iReadCount;
-        if( m_iReadCount > 0 )
-            return STATUS_SUCCESS;
-
-        if( m_iReadCount == 0 )
-        {
-            if( m_queue.empty() )
-                return STATUS_SUCCESS;
-
-            if( m_queue.front() == true )
-                return -EFAULT;
-
-            m_bWrite = true;
-            m_queue.pop_front();
-            return Sem_Post( &m_semWriter );
-        }
-        return -EFAULT;
-    }
-
-    inline gint32 ReleaseWrite()
-    {
-        CStdMutex oLock( m_oLock );
-        if( m_queue.empty() )
-        {
-            m_bWrite = false;
-        }
-        else if( m_queue.front() == false )
-        {
-            m_bWrite = true;
-            m_queue.pop_front();
-            Sem_Post( &m_semWriter );
-        }
-        else
-        {
-            m_bWrite = false;
-            while( m_queue.front() == true )
-            {
-                m_iReadCount++;
-                m_queue.pop_front();
-                Sem_Post( &m_semReader );
-                if( m_queue.empty() )
-                    break;
-            }
-        }
-        return STATUS_SUCCESS;
-    }
-
-    inline gint32 TryLockRead()
-    {
-        CStdMutex oLock( m_oLock );
-        if( !m_bWrite && m_queue.empty() )
-        {
-            m_iReadCount++;
-            oLock.Unlock();
-            return STATUS_SUCCESS;
-        }
-        return -EAGAIN;
-    }
-
-    inline gint32 TryLockWrite()
-    {
-        CStdMutex oLock( m_oLock );
-        if( m_bWrite || m_iReadCount > 0 ) 
-            return -EAGAIN;
-
-        if( m_iReadCount == 0 )
-        {
-            m_bWrite = true;
-            oLock.Unlock();
-            return 0;
-        }
-        return -EFAULT;
-    }
-
-};
-
 #define CFuseMutex  CStdRMutex
-
-template < bool bRead >
-struct CLocalLock
-{
-    // only works on the same thread
-    gint32 m_iStatus = STATUS_SUCCESS;
-    CSharedLock* m_pLock = nullptr;
-    bool m_bLocked = false;
-    CLocalLock( CSharedLock& oLock )
-    {
-        m_pLock = &oLock;
-        Lock();
-    }
-
-    ~CLocalLock()
-    { 
-        Unlock();
-        m_pLock = nullptr;
-    }
-
-    void Unlock()
-    {
-        if( !m_bLocked || m_pLock == nullptr )
-            return;
-        m_bLocked = false;
-        bRead ? m_pLock->ReleaseRead() :
-            m_pLock->ReleaseWrite();
-    }
-
-    void Lock()
-    {
-        if( !m_pLock )
-            return;
-        m_iStatus = ( bRead ?
-            m_pLock->LockRead() :
-            m_pLock->LockWrite() );
-        m_bLocked = true;
-    }
-
-    inline gint32 GetStatus() const
-    { return m_iStatus; }
-};
-
-typedef CLocalLock< true > CReadLock;
-typedef CLocalLock< false > CWriteLock;
 
 using DIR_SPTR=CHILD_TYPE;
