+++ include/propids.h
@@ -262,6 +262,7 @@ enum EnumPropId : gint32
     propFailureCount,   // type: a guint32 counter of the failed requests .
     propTaskAdded,      // type: a guint32 counter of the number of tasks added by the taskgroup
     propTaskRejected,   // type: a guint32 counter of the number of tasks rejected by the taskgroup
+    propStmHandle,      // type: a intptr as a handle of a stream. cannot pass machine boundary
 
     propReservedEnd = 0x10000000,
     propInvalid = -1,
